+++ Cmdline/Action/KSP.cs
@@ -1,4 +1,5 @@
 using System;
+using System.Collections.Generic;
 using System.IO;
 using System.Linq;
 using CKAN.Versioning;
@@ -54,8 +55,7 @@ namespace CKAN.CmdLine
                     {
                         // First the commands with three string arguments
                         case "fake":
-                            ht.AddPreOptionsLine($"Usage: ckan ksp {verb} [options] name path version dlcVersion");
-                            ht.AddPreOptionsLine($"Choose dlcVersion \"none\" if you want no simulated dlc.");
+                            ht.AddPreOptionsLine($"Usage: ckan ksp {verb} [options] name path version [--MakingHistory <version>] [--BreakingGround <version>]");
                             break;
 
                         case "clone":
@@ -124,8 +124,14 @@ namespace CKAN.CmdLine
             [ValueOption(0)] public string name { get; set; }
             [ValueOption(1)] public string path { get; set; }
             [ValueOption(2)] public string version { get; set; }
-            [ValueOption(3)] public string dlcVersion { get; set; }
-            [Option("set-default", DefaultValue = false, HelpText = "Set the new instance to the default one.")] public bool setToDefault { get; set; }
+
+            [Option("MakingHistory", DefaultValue = "none", HelpText = "The version of the Making History DLC to be faked.")]
+            public string makingHistoryVersion { get; set; }
+            [Option("BreakingGround", DefaultValue = "none", HelpText = "The version of the Breaking Ground DLC to be faked.")]
+            public string breakingGroundVersion { get; set; }
+
+            [Option("set-default", DefaultValue = false, HelpText = "Set the new instance as the default one.")]
+            public bool setDefault { get; set; }
         }
 
         // This is required by ISubCommand
@@ -303,8 +309,8 @@ namespace CKAN.CmdLine
 
             // Parse all options
             string instanceNameOrPath = options.nameOrPath;
-            string new_name = options.new_name;
-            string new_path = options.new_path;
+            string newName = options.new_name;
+            string newPath = options.new_path;
 
 
             log.Info("Cloning the KSP instance: " + options.nameOrPath);
@@ -320,16 +326,16 @@ namespace CKAN.CmdLine
                         if (instance.Name == instanceNameOrPath)
                         {
                             // Found it, now clone it.
-                            Manager.CloneInstance(instance, new_name, new_path);
+                            Manager.CloneInstance(instance, newName, newPath);
                             break;
                         }
                     }
                 }
                 // Try to use instanceNameOrPath as a path and create a new KSP object.
                 // If it's valid, go on.
-                else if (new CKAN.KSP(instanceNameOrPath, new_name, User) is CKAN.KSP instance && instance.Valid)
+                else if (new CKAN.KSP(instanceNameOrPath, newName, User) is CKAN.KSP instance && instance.Valid)
                 {
-                    Manager.CloneInstance(instance, new_name, new_path);
+                    Manager.CloneInstance(instance, newName, newPath);
 
                 }
                 // There is no instance with this name or at this path.
@@ -370,12 +376,16 @@ namespace CKAN.CmdLine
                 ListInstalls();
                 return Exit.ERROR;
             }
-
+            catch (InstanceNameTakenKraken kraken)
+            {
+                User.RaiseError("This instance name is already taken: {0}", kraken.instName);
+                return Exit.BADOPT;
+            }
 
             // Test if the instance was added to the registry.
             // No need to test if valid, because this is done in AddInstance(),
             // so if something went wrong, HasInstance is false.
-            if (Manager.HasInstance(new_name))
+            if (Manager.HasInstance(newName))
             {
                 return Exit.OK;
             }
@@ -506,10 +516,26 @@ namespace CKAN.CmdLine
         /// </summary>
         private int FakeNewKSPInstall (FakeOptions options)
         {
+            int error()
+            {
+                log.Debug("Instance faking failed, see console output for details.");
+                User.RaiseMessage("--Error--");
+                return Exit.ERROR;
+            }
+
+            int badArgument()
+            {
+                log.Debug("Instance faking failed: bad argument(s). See console output for details.");
+                User.RaiseMessage("--Error: bad argument(s)--");
+                return Exit.BADOPT;
+            }
+
+
             if (options.name == null || options.path == null || options.version == null)
             {
-                User.RaiseMessage("ksp fake <name> <path> <version> [dlcVersion] - argument(s) missing");
-                return Exit.BADOPT;
+                User.RaiseMessage("ksp fake <name> <path> <version> " +
+                	"[--MakingHistory <version>] [--BreakingGround <version>] - argument(s) missing");
+                return badArgument();
             }
 
             log.Debug("Parsing arguments...");
@@ -517,16 +543,33 @@ namespace CKAN.CmdLine
             string installName = options.name;
             string path = options.path;
             KspVersion version;
-            bool setToDefault = options.setToDefault;
-            // dlcVersion is null if a user wants no simulated DLC
-            string dlcVersion;
-            if (options.dlcVersion == null || options.dlcVersion.ToLower() == "none")
+            bool setDefault = options.setDefault;
+
+            // options.<DLC>Version is "none" if the DLC should not be simulated.
+            Dictionary<DLC.IDlcDetector, KspVersion> dlcs = new Dictionary<DLC.IDlcDetector, KspVersion>();
+            if (options.makingHistoryVersion != null && options.makingHistoryVersion.ToLower() != "none")
             {
-                dlcVersion = null;
+                if (KspVersion.TryParse(options.makingHistoryVersion, out KspVersion ver))
+                {
+                    dlcs.Add(new DLC.MakingHistoryDlcDetector(), ver);
+                }
+                else
+                {
+                    User.RaiseError("Please check the Making History DLC version argument - Format it like Maj.Min.Patch - e.g. 1.1.0");
+                    return badArgument();
+                }
             }
-            else
+            if (options.breakingGroundVersion != null && options.breakingGroundVersion.ToLower() != "none")
             {
-                dlcVersion = options.dlcVersion;
+                if (KspVersion.TryParse(options.breakingGroundVersion, out KspVersion ver))
+                {
+                    dlcs.Add(new DLC.BreakingGroundDlcDetector(), ver);
+                }
+                else
+                {
+                    User.RaiseError("Please check the Breaking Ground DLC version argument - Format it like Maj.Min.Patch - e.g. 1.1.0");
+                    return badArgument();
+                }
             }
 
             // Parse the choosen KSP version
@@ -538,7 +581,7 @@ namespace CKAN.CmdLine
             {
                 // Thrown if there is anything besides numbers and points in the version string or a different syntactic error.
                 User.RaiseError("Please check the version argument - Format it like Maj.Min.Patch[.Build] - e.g. 1.6.0 or 1.2.2.1622");
-                return Exit.BADOPT;
+                return badArgument();
             }
 
             // Get the full version including build number.
@@ -546,41 +589,54 @@ namespace CKAN.CmdLine
             {
                 version = version.RaiseVersionSelectionDialog(User);
             }
-            catch (IncorrectKSPVersionKraken)
+            catch (BadKSPVersionKraken)
             {
                 User.RaiseError("Couldn't find a valid KSP version for your input.\n" +
                 	"Make sure to enter the at least the version major and minor values in the form Maj.Min - e.g. 1.5");
-                return Exit.BADOPT;
+                return badArgument();
             }
             catch (CancelledActionKraken)
             {
                 User.RaiseError("Selection cancelled! Please call 'ckan ksp fake' again.");
-                return Exit.ERROR;
+                return error();
             }
 
-
             User.RaiseMessage(String.Format("Creating new fake KSP install {0} at {1} with version {2}", installName, path, version.ToString()));
+            log.Debug("Faking instance...");
 
             try
             {
                 // Pass all arguments to CKAN.KSPManager.FakeInstance() and create a new one.
-                Manager.FakeInstance(installName, path, version, dlcVersion);
-                if (setToDefault)
+                Manager.FakeInstance(installName, path, version, dlcs);
+                if (setDefault)
+                {
                     User.RaiseMessage("Setting new instance to default...");
                     Manager.SetAutoStart(installName);
+                }
+            }
+            catch (InstanceNameTakenKraken kraken)
+            {
+                User.RaiseError("This instance name is already taken: {0}", kraken.instName);
+                return badArgument();
             }
             catch (BadInstallLocationKraken kraken)
             {
                 // The folder exists and is not empty.
-                log.Error(kraken);
-                return Exit.ERROR;
+                User.RaiseError(kraken.Message);
+                return badArgument();
+            }
+            catch (WrongKSPVersionKraken kraken)
+            {
+                // Thrown because the specified KSP version is too old for one of the selected DLCs.
+                User.RaiseError(kraken.Message);
+                return badArgument();
             }
             catch (NotKSPDirKraken kraken)
             {
                 // Something went wrong adding the new instance to the registry,
-                // most likely because the newly created directory is not valid.
+                // most likely because the newly created directory is somehow not valid.
                 log.Error(kraken);
-                return Exit.ERROR;
+                return error();
             }
             catch (InvalidKSPInstanceKraken)
             {
@@ -600,7 +656,7 @@ namespace CKAN.CmdLine
             {
                 User.RaiseError("Something went wrong. Try to add the instance yourself with \"ckan ksp add\".",
                     "Also look if the new directory has been created.");
-                return Exit.ERROR;
+                return error();
             }
         }
         #endregion